fix coverage bugs

+++ test/legacy_test/test_gather_op.py
@@ -428,7 +428,7 @@ class API_TestGather(unittest.TestCase):
             place = base.CPUPlace()
             exe = base.Executor(place)
             input = np.array([[1, 2], [3, 4], [5, 6]]).astype('float64')
-            index_1 = np.array([1, 2])
+            index_1 = np.array([1, 2]).astype('int64')
             (result,) = exe.run(
                 feed={"data1": input, "index": index_1}, fetch_list=[out]
             )
